refactor: restructure comments widget to nested Comments component
