+++ packages/frontend/src/app/events/[event_id]/page.tsx
@@ -4,7 +4,7 @@ import { GET_EVENTS_BY_ID } from '@/app/graphql/queries/events';
 import { Event as EventType } from '@/app/graphql/types';
 import Spinner from '@/assets/icons/spinner.svg';
 import { Backdrop } from '@/components/widgets/Backdrop';
-import { CommentsList } from '@/components/widgets/CommentsList';
+import { CommentsList } from '@/components/widgets/Comments';
 import { Event } from '@/components/widgets/Event';
 import { useQuery } from '@apollo/client/react';
 import type { NextPage } from 'next';
