+++ packages/frontend/src/components/widgets/Comments/CommentForm/CommentForm.module.css
@@ -1,0 +1,8 @@
+.form {
+    margin-bottom: 0.75rem;
+    width: 50rem;
+}
+.form > *:not(:last-child) {
+    display: block;
+    margin: 1rem 0;
+}
