+++ packages/frontend/src/components/widgets/Comments/Comments.module.css
@@ -1,0 +1,65 @@
+.container {
+    width: 50rem;
+}
+
+.title {
+    margin-bottom: 2.75rem;
+    font-size: 1.25rem;
+    line-height: 2rem;
+    font-weight: 700;
+    color: var(--main-color);
+}
+
+.replies {
+    margin-left: 6rem;
+}
+
+.messageContainer {
+    padding: 1rem 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.messageContainer > svg {
+    width: 3rem;
+    height: 3rem;
+}
+
+.error {
+    font-size: 1.3rem;
+    font-weight: 600;
+    color: var(--error-color);
+}
+
+.commentsList {
+    margin-bottom: 1.5rem;
+}
+
+.loadButton {
+    border: none;
+    text-transform: uppercase;
+    font-size: 1rem;
+}
+
+.buttonText {
+    color: var(--additional-color);
+    font-family: inherit;
+    font-size: 1rem;
+    font-weight: 700;
+    line-height: 1.19rem;
+    text-transform: uppercase;
+}
+
+.buttonText:hover {
+    color: var(--main-color);
+}
+
+.buttonActive {
+    color: var(--main-color);
+    font-family: inherit;
+    font-size: 1rem;
+    font-weight: 700;
+    line-height: 1.19rem;
+    text-transform: uppercase;
+}
