+++ packages/frontend/src/components/widgets/Comments/index.ts
@@ -1,0 +1,1 @@
+export { Comments } from './Comments';
