+++ packages/frontend/src/components/widgets/Comments/types.ts
@@ -1,0 +1,4 @@
+export interface ReplyTo {
+    id: string;
+    userName: string;
+}
