--- packages/frontend/src/components/widgets/CommentsList/CommentsList.module.css
@@ -1,43 +1,0 @@
-.container {
-  width: 50rem;
-}
-
-.title {
-  margin-bottom: 2.75rem;
-  font-size: 1.25rem;
-  line-height: 2rem;
-  font-weight: 700;
-  color: var(--main-color);
-}
-
-.replies {
-  margin-left: 6rem;
-}
-
-.messageContainer {
-  padding: 1rem 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.messageContainer > svg {
-  width: 3rem;
-  height: 3rem;
-}
-
-.error {
-  font-size: 1.3rem;
-  font-weight: 600;
-  color: var(--error-color);
-}
-
-.commentsList {
-  margin-bottom: 1.5rem;
-}
-
-.loadButton {
-  border: none;
-  text-transform: uppercase;
-  font-size: 1rem;
-}
