--- packages/frontend/src/components/widgets/CommentsList/index.ts
@@ -1,1 +1,0 @@
-export { CommentsList } from './CommentsList';
