--- packages/frontend/src/components/widgets/CommentsList/types.ts
@@ -1,11 +1,0 @@
-export interface ReplyTo {
-    id: string;
-    userName: string;
-}
-
-export interface ICommentAuthor {
-    _id: string;
-    firstName: string;
-    lastName: string;
-    image?: string | null;
-}
